The version of angular-ui-router wasn't compatible with angulartics.
The 1st symptom  was actually felt a couple of weeks ago. I never
managed to keep track of the regular angular-ui events described in
`https://github.com/angular-ui/ui-router/wiki`.

Reading the source code
https://github.com/angulartics/angulartics/blob/master/src/angulartics.j
s#L218 I noticed that if the event tracking didn't work, then the
analytics tracking wouldn't either.

By enabling the debugging mode on angulartics, I was able to make sure
that no event was being tracked and that after downgrading
angular-ui-router, it was back to the game.